src/e/forms/FormDialog.java: src/e/gui/EHistoryComboBoxModel.java: src/e/gui/EPopupMenu.java: src/e/gui/FilteredListModel.java: src/e/ptextarea/PAnchorSet.java: src/e/ptextarea/PCLikeTextStyler.java: src/e/ptextarea/PTextArea.java: src/e/ptextarea/PTextAreaSpellingChecker.java: src/e/toys/world/Gazetteer.java: src/e/util/InstanceTracker.java: src/e/util/NumberDecoder.java: (bug 39) remove unnecessary casts, make some types more specific, and make more use of the new for statement.

+++ salma-hayek/trunk/src/e/gui/FilteredListModel.java
@@ -56,7 +56,7 @@ public class FilteredListModel extends AbstractListModel {
     }
     
     public Object getElementAt(int index) {
-        int newIndex = ((Integer) validIndexes.get(index)).intValue();
+        int newIndex = validIndexes.get(index);
         return model.getElementAt(newIndex);
     }
